F/F: fixed and improved transplanted functions especially JWT auth

+++ src/users/users.controller.ts
@@ -1,11 +1,13 @@
 import {
-  Body, Controller, DefaultValuePipe, Delete, Get, Headers, HttpCode, NotFoundException, Param, ParseIntPipe, ParseUUIDPipe, Post, Query, Res, UseGuards,
+  Body, Controller, Delete, ForbiddenException, Get, HttpCode, NotFoundException, Param, ParseUUIDPipe, Post, Request, UseGuards,
 } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UserLoginDto } from './dto/user-login.dto';
 import { User } from './dto/user-info.dto';
 import { UsersService } from './users.service';
 import { AuthService } from 'src/auth/auth.service';
+import { AuthGuard } from 'src/auth/auth.guard';
+import { RequestWithUser } from 'src/common/user.interface';
 
 // [TODO] add service logic and res for each endpoints
 @Controller('users')
@@ -14,14 +16,6 @@ export class UsersController {
     private readonly usersService: UsersService,
     private readonly authService: AuthService,
   ) { }
-
-  @Get()
-  findAll(
-    @Query('offset', new DefaultValuePipe(0)) offset: number,
-    @Query('limit', new DefaultValuePipe(10), ParseIntPipe) limit: number,
-  ) {
-    return;
-  }
 
   @HttpCode(201)
   @Post()
@@ -40,22 +34,33 @@ export class UsersController {
     return res;
   }
 
+  @UseGuards(AuthGuard)
   @Get('/:userUuid')
-  async getUserInfo(@Param('userUuid', new ParseUUIDPipe({ errorHttpStatusCode: 406 })) userUuid: string,
-    @Headers() headers: Headers & { Authorization: string }): Promise<User> {
-    const { Authorization: token } = headers;
-    this.authService.verify(token);
+  async getUserInfo(
+    @Param('userUuid', new ParseUUIDPipe({ errorHttpStatusCode: 406 })) userUuid: string,
+    @Request() req: RequestWithUser,
+  ): Promise<User> {
+    const userAuthInfo = req.payload;
+    if (userAuthInfo.userUuid !== userUuid) {
+      throw new ForbiddenException('cannot access to other user info');
+    }
 
     const userInfo = await this.usersService.getUserInfo(userUuid);
-
     return userInfo;
   }
 
+  @UseGuards(AuthGuard)
   @HttpCode(204)
   @Delete('/:userUuid')
   async remove(
     @Param('userUuid', new ParseUUIDPipe({ errorHttpStatusCode: 406 })) userUuid: string,
+    @Request() req: RequestWithUser,
   ) {
+    const userAuthInfo = req.payload;
+    if (userAuthInfo.userUuid !== userUuid) {
+      throw new ForbiddenException('cannot access to other user info');
+    }
+
     await this.usersService.remove(userUuid).then((isRemoved) => {
       if (!isRemoved) {
         throw new NotFoundException('user not found');
